Renamed starfish to starfix and Windows installer fixed
