+++ .github/workflows/release-build-native-linux.yml
@@ -39,8 +39,8 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         upload_url: ${{ github.event.release.upload_url }}
-        asset_path: ./cli/target/starfish-${{ github.event.release.tag_name }}.zip
-        asset_name: starfish-linux.zip
+        asset_path: ./cli/target/starfix-${{ github.event.release.tag_name }}.zip
+        asset_name: starfix-linux.zip
         asset_content_type: application/octet-stream
 
 # Trigger this job only on Push or Pull-Request
@@ -64,8 +64,8 @@ jobs:
         id: upload-native-executable
         uses: actions/upload-artifact@v2
         with:
-          name: starfish-linux-${{github.sha}}.zip
-          path: ./cli/target/starfish-${{github.sha}}.zip
+          name: starfix-linux-${{github.sha}}.zip
+          path: ./cli/target/starfix-${{github.sha}}.zip
