+++ .github/workflows/release-build-native-win64.yml
@@ -50,8 +50,8 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         upload_url: ${{ github.event.release.upload_url }}
-        asset_path: ./cli/target/starfish-${{ github.event.release.tag_name }}.zip
-        asset_name: starfish-win64.zip
+        asset_path: ./cli/target/starfix-${{ github.event.release.tag_name }}.zip
+        asset_name: starfix-win64.zip
         asset_content_type: application/octet-stream
 
 # Trigger this job only on Push or Pull-Request
@@ -86,5 +86,5 @@ jobs:
         id: upload-native-executable
         uses: actions/upload-artifact@v2
         with:
-          name: starfish-win64-${{github.sha}}.zip
-          path: ./cli/target/starfish-${{github.sha}}.zip
+          name: starfix-win64-${{github.sha}}.zip
+          path: ./cli/target/starfix-${{github.sha}}.zip
