+++ README.adoc
@@ -1,8 +1,8 @@
-# Project Starfish 
+# Project Starfix
 
 A  cross platform client side application that would let you open a file in IDE of your choice (vscode, eclipse, intellij, emacs, vi, etc.) and  other cammands locally directly from the browser or file system.This will enable an option to “Open in IDE” through browser extension on various websites like github, gitlab,gitbucket etc similar to one you see in this demo below.
 
-image::https://user-images.githubusercontent.com/31308705/79685056-61078280-8253-11ea-8ac1-aab1531ca0ab.gif[Starfish Demo]
+image::https://user-images.githubusercontent.com/31308705/79685056-61078280-8253-11ea-8ac1-aab1531ca0ab.gif[Starfix Demo]
 
 ## Workflow
 - Click on "Open in IDE" on Repository page.
@@ -34,15 +34,15 @@ image::https://user-images.githubusercontent.com/31308705/79685056-61078280-8253
 - Building native executable: `mvnw package -Dnative`
 
 ## Useful resources:
-As Starfish uses Command Mode in Quarkus refer these links to learn more about it:
+As Starfix uses Command Mode in Quarkus refer these links to learn more about it:
 
 - https://quarkus.io/blog/introducing-command-mode/[Introducing Command Mode ] written by https://github.com/maxandersen[@maxandersen]
 - https://quarkus.io/guides/command-mode-reference[QUARKUS - COMMAND MODE APPLICATIONS]
 
-## Contributing to Starfish
-- Issues: Start by looking at the existing issues at https://github.com/maxandersen/starfish/issues[starfish/issues] and create a Pull Request with appropriate description/comment.
-- Bugs: Run the starfish application and if you find unexpected behaviour then open a new issue with label `bug` and include a proper description/comment.
-- Feature request: Brainstorm ideas and bring them to surface on the features that can be implemented in starfish and then create an issue with label as `enahancement`.Always include as exhaustive description as possible to help understand better.
+## Contributing to Starfix
+- Issues: Start by looking at the existing issues at https://github.com/starfixdev/starfix/issues[Starfix/issues] and create a Pull Request with appropriate description/comment.
+- Bugs: Run the Starfix application and if you find unexpected behaviour then open a new issue with label `bug` and include a proper description/comment.
+- Feature request: Brainstorm ideas and bring them to surface on the features that can be implemented in Starfix and then create an issue with label as `enahancement`.Always include as exhaustive description as possible to help understand better.
 
 ## Technology for Development
 - https://quarkus.io/[Quarkus]
