+++ cli/installer/windows/installer/Starfix/starfix.cmd
@@ -1,0 +1,3 @@
+@echo off
+cd "C:\Users\Public\Starfix"
+starfix.exe %1
