+++ cli/installer/windows/installer/install.cmd
@@ -1,7 +1,7 @@
 @echo off
 cd %~dp0
-xcopy /s/i "%~dp0\Starfish" "C:\Users\Public\Starfish"
+xcopy /s/i "%~dp0\Starfix" "C:\Users\Public\Starfix"
 echo Please accept the upcoming Prompts to install successfully
-"%~dp0\starfish.reg" 
-setx path "%path%;C:\Users\Public\Starfish"
+"%~dp0\starfix.reg" 
+setx path "%path%;C:\Users\Public\Starfix"
 PAUSE
