--- cli/installer/windows/installer/starfish.cmd
@@ -1,3 +1,0 @@
-@echo off
-cd "C:\Users\Public\Starfish"
-starfish.exe %1
