+++ cli/src/main/java/dev/starfix/starfix/JavaMain.java
@@ -1,4 +1,4 @@
-package dk.xam.starfish;
+package dev.starfix.starfix;
 import io.quarkus.runtime.Quarkus;
 import io.quarkus.runtime.annotations.QuarkusMain;
 
@@ -6,6 +6,6 @@ import io.quarkus.runtime.annotations.QuarkusMain;
 public class JavaMain {
 
     public static void main(String... args) {
-        Quarkus.run(Starfish.class, args);
+        Quarkus.run(Starfix.class, args);
     }
 }
