+++ cli/src/test/java/dk/xam/StarfixTestIT.java
@@ -1,9 +1,9 @@
-package dk.xam;
+package dev.starfix;
 
 import io.quarkus.test.junit.NativeImageTest;
 
 @NativeImageTest
-public class StarfishTestIT extends StarfishTest {
+public class StarfixTestIT extends StarfixTest {
 
     // Execute the same tests but in native mode.
 }
